Add option to specify the database URL
- The option -d <dbUri> is used to specify the address of a mysql database. It also accepts admin username and password for the database

- The DBHelper tries to load the db.pwd. If it fails and a URI is specified as argument of the setUpDB method, it will look for username and password in the URI.
Moreover, if a URI is specified, that URI will be validated before use. On the contrary, if no URI is specified, the default address (jdbc:mysql://localhost:3306) is used

+++ AceAS/src/main/java/eu/sifishome/DBHelper.java
@@ -11,6 +11,8 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.sql.SQLException;
 
 /**
@@ -32,6 +34,7 @@ public class DBHelper
 
     private static String dbAdminUser = null;
     private static String dbAdminPwd = null;
+    private static String dbUrl;
 
     /**
      * Sets up the DB using the current default adapter.
@@ -39,13 +42,33 @@ public class DBHelper
      * @throws AceException if acting on the database fails
      * @throws IOException if loading admin information fails
      */
-    public static void setUpDB() throws AceException, IOException
+    public static void setUpDB(String dbUrl) throws AceException, IOException
     {
         // First load the DB admin username and password from an external file.
-        loadAdminLoginInformation();
-        
+        try {
+            loadAdminLoginInformation();
+        } catch (IOException e) {
+            // if the dbUrl is not empty, later we try to load admin username and password from the URL
+            if (dbUrl == null) {
+                throw new IOException(e.getMessage());
+            }
+        }
+
+        // Parse the DB url.
+        // If the url includes db credentials, these will be used,
+        // and they possibly override those loaded from the db.pwd file
+        try {
+            parseDbUrl(dbUrl);
+        } catch (URISyntaxException e) {
+            throw new AceException(e.getMessage());
+        }
+
+        if (dbAdminUser == null || dbAdminPwd == null) {
+            throw new AceException("Cannot retrieve admin username and password for the database");
+        }
+
         // Set parameters for the DB.
-        dbAdapter.setParams(testUsername, testPassword, testDBName, null);
+        dbAdapter.setParams(testUsername, testPassword, testDBName, DBHelper.dbUrl);
 
         // In case database and/or user already existed.
         SQLConnector.wipeDatabase(dbAdapter, dbAdminUser, dbAdminPwd);
@@ -111,5 +134,85 @@ public class DBHelper
                 line = br.readLine();
             }
         }
+    }
+
+    /**
+     * Parse the database url provided as input.
+     * This method returns null if the string passed as argument is null.
+     * This method accepts only urls that start with "jdbc:mysql://" and ignores
+     * the path of the url, if any.
+     * It returns a string in the form "jdbc:mysql://host:port". If the port is not
+     * specified in the provided url, it places the default port (3306) in the
+     * returned string
+     *
+     * @param url the url of the database
+     * @throws URISyntaxException If the given string violates RFC 2396
+     */
+    private static void parseDbUrl(String url) throws URISyntaxException, AceException {
+        if (url == null) {
+            // using default db Url
+            return;
+        }
+        if (!url.startsWith("jdbc:mysql://")) {
+            throw new IllegalArgumentException("Wrong database URL. The URL must start with \"jdbc:mysql://\"");
+        }
+        //strip off the jdbc: part
+        url = url.substring(5);
+
+        URI uri = new URI(url);
+        String host = uri.getHost();
+        int port = uri.getPort();
+
+        if (host == null) {
+            throw new AceException("Wrong database URL. The host cannot be parsed");
+        }
+
+        // if the port is not defined, use the default port (3306)
+        if (port != -1) {
+            dbUrl = "jdbc:mysql://" + host + ":" + port;
+        }
+        else {
+            dbUrl = "jdbc:mysql://" + host + ":3306";
+        }
+
+        overrideCredentials(uri);
+    }
+
+    /**
+     * Set admin username and password, if provided.
+     * This method possibly overrides the admin credentials provided within the db.pwd file.
+     *
+     * @param uri the URI
+     * @throws AceException if some error occurs during parsing of the URI
+     */
+    private static void overrideCredentials(URI uri) throws AceException {
+
+        String credentials = uri.getUserInfo();
+        if (credentials == null) {
+            return;
+        }
+        String[] splitCredentials = credentials.split(":");
+        if (splitCredentials.length > 2) {
+            throw new AceException("Wrong database URL. User info cannot be parsed. Too many colons");
+        }
+
+        if (splitCredentials[0].equals("")) {
+            throw new AceException("Wrong database URL. Username cannot be empty");
+        }
+        dbAdminUser = splitCredentials[0];
+
+        // if password is present
+        if (splitCredentials.length == 2) {
+            dbAdminPwd = splitCredentials[1];
+        }
+        else {
+            dbAdminPwd = "";
+        }
+        if (dbAdminPwd.equals("")) {
+            System.out.println("Warning: no password for user " + dbAdminUser + " was specified. " +
+                    "Using an empty password..." );
+        }
+
+        System.out.println("Credentials loaded from the provided database URI");
     }
 }
